+++ builtin/plugins/dposv3/dpos.go
@@ -166,10 +166,6 @@ type (
 	GetRequestBatchTallyRequest = dtypes.GetRequestBatchTallyRequest
 )
 
-type DPOSTotalStaked struct {
-	TotalStaked *types.BigUInt
-}
-
 type DPOS struct {
 }
