+++ config/config.go
@@ -9,9 +9,9 @@ import (
 	"os"
 	"path"
 	"path/filepath"
+	"strings"
 
 	"github.com/loomnetwork/loomchain/auth"
-	"github.com/loomnetwork/loomchain/builtin/plugins/dposv3"
 	plasmacfg "github.com/loomnetwork/loomchain/builtin/plugins/plasma_cash/config"
 	genesiscfg "github.com/loomnetwork/loomchain/config/genesis"
 	"github.com/loomnetwork/loomchain/events"
@@ -151,7 +151,7 @@ type Config struct {
 
 	Web3 *eth.Web3Config
 
-	DPOS *dposv3.DPOSConfig
+	DPOS *DPOSConfig
 }
 
 type Metrics struct {
@@ -170,6 +170,33 @@ func DefaultFnConsensusConfig() *FnConsensusConfig {
 		Enabled: false,
 		Reactor: fnConsensus.DefaultReactorConfigParsable(),
 	}
+}
+
+type DPOSConfig struct {
+	BootstrapNodes           []string
+	TotalStakedCacheDuration int64
+}
+
+func DefaultDPOSConfig() *DPOSConfig {
+	return &DPOSConfig{
+		BootstrapNodes: []string{
+			"default:0x0e99fc16e32e568971908f2ce54b967a42663a26",
+			"default:0xac3211caecc45940a6d2ba006ca465a647d8464f",
+			"default:0x69c48768dbac492908161be787b7a5658192df35",
+			"default:0x2a3a7c850586d4f80a12ac1952f88b1b69ef48e1",
+			"default:0x4a1b8b15e50ce63cc6f65603ea79be09206cae70",
+			"default:0x0ce7b61c97a6d5083356f115288f9266553e191e",
+		},
+		TotalStakedCacheDuration: 60, // 60 seconds
+	}
+}
+
+func (dposCfg *DPOSConfig) BootstrapNodesList() map[string]bool {
+	bootstrapNodesList := map[string]bool{}
+	for _, addr := range dposCfg.BootstrapNodes {
+		bootstrapNodesList[strings.ToLower(addr)] = true
+	}
+	return bootstrapNodesList
 }
 
 type DBBackendConfig struct {
@@ -427,7 +454,7 @@ func DefaultConfig() *Config {
 	cfg.EventStore = events.DefaultEventStoreConfig()
 	cfg.EvmStore = evm.DefaultEvmStoreConfig()
 	cfg.Web3 = eth.DefaultWeb3Config()
-	cfg.DPOS = dposv3.DefaultDPOSConfig()
+	cfg.DPOS = DefaultDPOSConfig()
 
 	cfg.FnConsensus = DefaultFnConsensusConfig()
 
@@ -802,7 +829,7 @@ DPOS:
   {{- range .DPOS.BootstrapNodes}}
     - "{{. -}}"
   {{- end}}
-  # Time duration in seconds in which TotolStaked is valid  
+  # How long (in seconds) the response from the dpos_total_staked RPC method should be cached.
   TotalStakedCacheDuration: {{ .DPOS.TotalStakedCacheDuration }}
 {{end}}
