+++ rpc/mock_query_server.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/loomnetwork/go-loom/plugin/types"
 
-	"github.com/loomnetwork/loomchain/builtin/plugins/dposv3"
 	"github.com/loomnetwork/loomchain/config"
 	"github.com/loomnetwork/loomchain/rpc/eth"
 	"github.com/loomnetwork/loomchain/vm"
@@ -271,7 +270,7 @@ func (m *MockQueryService) GetContractRecord(addr string) (*types.ContractRecord
 	return nil, nil
 }
 
-func (m *MockQueryService) DPOSTotalStaked() (*dposv3.DPOSTotalStaked, error) {
+func (m *MockQueryService) DPOSTotalStaked() (*DPOSTotalStaked, error) {
 	m.MethodsCalled = append([]string{"DposTotalStaked"}, m.MethodsCalled...)
 	return nil, nil
 }
